+++ CHANGELOG.md
@@ -15,3 +15,4 @@
 - removed test step in wercker.yml (conor.fennell)
 - refactored javascript code on UI (brian.lough)
 - added product info for a stock check page (brian.lough)
+- added basic stock check page (brian.lough)